refactor: - 1급 컬렉션 객체 생성 (Sessions, Students) - 수강 신청 로직을 session 객체 내에서 하도록 registerStudent 메서드 추가 - Course 도메인에 sessions 추가 - Session 도메인에 Students 추가, Student 도메인에 Sessions 추가

+++ src/main/java/nextstep/courses/infrastructure/JdbcSessionRepository.java
@@ -39,7 +39,7 @@ public class JdbcSessionRepository implements SessionRepository {
                     SessionStatus.valueOf(rs.getString("status")),
                     SessionType.valueOf(rs.getString("type")),
                     rs.getInt("capacity"),
-                    rs.getInt("fee")
+                    rs.getLong("fee")
             );
         }, id);
     }
